requirements.txt
django==4.2.7
torch>=2.6.0
transformers>=4.36.2
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
numpy>=1.24.3
pandas>=2.0.3
python-dotenv==1.0.0
PyYAML==6.0.1
scikit-learn>=1.3.2
tqdm>=4.66.1
gunicorn==21.2.0
